support diagram direction
